functions/requirements.txt
azure-functions
pymongo
